Create contacts module, ditch custom config file.
In hindsight, we expect any users of this code to edit it for site-specific
needs anyway - it's not a library, it's an 'app' - so it makes sense to store
any contact details in the version control along with the code.

In the public code, we simply provide a template file as an example.